Infra: Jesus Christ these scripts I swear to God
Why does anyone inflict this on themselves.

package.json scripts is a piece of shit and it is bad and it should feel
bad. Holy hell, this is the most broken-ass task management tool.

Yes, let's just put scripts in strings and add a bunch of stupid
variables to the environment and just pipe all that hot garbage into
bash.

Kudos to bash. Yeah, you can really hand bash a bag of shit -- it'll
still try to find the pearls among the feces trying, hard as it can, to
fight off the swine. What a wingman.

Let's top one off to bash tonight, shall we?

By the way: npm, get the fuck out of my build. The next time I do this
I'm using goddamn make. That's how bad you are.